Conflicts:
	src/game/AntiCheat.cpp
	src/game/Level3.cpp
	src/game/ObjectMgr.cpp
	src/game/ObjectMgr.h
	src/game/Player.cpp
	src/game/Spell.cpp
	src/game/WardenHandler.cpp
	src/game/WardenHandler.h

+++ src/game/Level3.cpp
@@ -5067,8 +5067,8 @@ bool ChatHandler::HandleCharacterLevelCommand(char* args)
     if (newlevel > STRONG_MAX_LEVEL)                        // hardcoded maximum level
         newlevel = STRONG_MAX_LEVEL;
 
-    if ((uint32)newlevel > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-        newlevel = sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL);
+    if (newlevel > int32(sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL)))
+        newlevel = int32(sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL));
 
     HandleCharacterLevel(target, target_guid, oldlevel, newlevel);
 
@@ -5115,8 +5115,8 @@ bool ChatHandler::HandleLevelUpCommand(char* args)
     if (newlevel > STRONG_MAX_LEVEL)                        // hardcoded maximum level
         newlevel = STRONG_MAX_LEVEL;
 
-    if ((uint32)newlevel > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-        newlevel = sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL);
+    if (newlevel > int32(sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL)))
+        newlevel = int32(sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL));
 
     HandleCharacterLevel(target, target_guid, oldlevel, newlevel);
